offer vault decryption dialog if the autofill vault is still locked during the autofill request
