+++ app/src/main/AndroidManifest.xml
@@ -42,10 +42,13 @@
         <activity
             android:name=".activities.ShortcutActivity"
             android:theme="@android:style/Theme.NoDisplay" />
+        <activity
+            android:name=".activities.AutofillInteractionActivity"
+            android:theme="@style/AppTheme.NoActionBar"/>
 
         <service
             android:name=".autofill.CredentialAutofillService"
-            android:exported="true"
+            android:exported="false"
             android:label="Passman Credential Autofill Service"
             android:permission="android.permission.BIND_AUTOFILL_SERVICE">
             <meta-data
